CTS: Remove description generator shell script and related cleanup
The description generator shell script has been replaced by a Python script
that has fewer dependencies on the operating system. This change removes the
old shell script.
The doclet used to insert an XML_INFO attribute into the generated description.
The shell script used sed to replace this with any additional attributes it
wanted to inject. With the new Python script, this extra attribute is no longer
required.

Change-Id: I2aadf2dea55b5f263406807e76221476259b48ae