Add Feature List to CTS Report Header
Bug 2742459

Use the PackageManager to check which features the device has
enabled. Add them to the XML in separate block like
<FeatureInfo><Feature... /></FeatureInfo> and display them in
the report with ascii check boxes like "[X] android.hardware.camera".

In the future the list of features could be gleamed from
the XML files in frameworks/base/data/etc...

Change-Id: Ife7aa1fdfa060aecbc3438eae3f7b33f70a91f80