Fix Missing CTS Test Cases
CTS uses JavaDoc to scan the source code tree for tests. Since
libcore moved to a separate project, javadoc started to drop all
the testcases that referenced JUnit or the Dalvik annotations
leading to CTS dropping most of its test cases.

It would be nice if JavaDoc failed more loudly.

Change-Id: I0766e31212732c5d05d921790c849fc0e95db61e