Report Installed Apps and Properties
Bug 5875919

Add a list of installed packages and properties to the report.
There are a lot of packages and properties, so I restructured the
device information section to be four columns wide and split apart
the test summary into its own area. The device information is
no longer shown by default, since it is so long and wide. However,
there is now a link at the top that allows you to toggle it.

Refactor DeviceInfoResult to utilize common functions for
serializing and parsing. Make sure it all works by adding tests
for it in DeviceInfoResultTest. The tests weren't running,
because they were not part of UnitTests. Finally, fix the
existing CtsXmlResultReporterTest failures. The mock was not
being setup properly.

Change-Id: If8f909444785267806d1e52f5270146962255271