Remove Most of dx-tests
Strip out the dx-tests that no longer compile. The build doesn't
break, because the compilation occurs in a script. These tests
haven't been running as part of CTS and are not maintained at all.

Keep the cfassembler module since that is needed by the vm-tests
which are still maintained. Rename the dx-tests directory to
cfassembler, since that is all that is in it now.

Change-Id: I038054ad0703bdd00af3b7b1012ea9ad3eabd8a6