Move accessibility CTS tests to UiAutomation and fixing all broken ones.
Accessibility tests were using a delegating accessibility service that
is reinstalled and enabled by the shell test driver since running
instrumentation restarts the tested package which breaks the binding
between the system and an accessibility service. Now that we have public
UI testing APIs, this ugliness is not needed. This change moves the
relevant accessibility tests to use the UiAutomationa APIs.

bug:8695422
bug:8657560

Change-Id: I7130d91aeea32f86548f46ce24bd04a198700d36