Revise test plans for continuous test runs
Updated continuous test plans - usually used to run daily
tests during the development of the device or platform, and
running on non release builds - so that it does not include
tests expected to run on user/release-key builds for final
Compatibility testing.

Also updated a bug in the continuous test plans where there
was some redundancy (e.g. deqp test)

Bug: 20295773
Change-Id: I2063fb546bbdb3c16dd06d1d9de91825296b40e9