Use general PropertyCheck class for preconditions
Check the device's build type and locale with PropertyCheck, that checks
any property accessible from the ITestDevice.getProperty()
method. Notable changes in functionality include the option to print a
warning instead of throwing an exception if the device's locale isn't as
expected, and the removal of a default expected locale, which must now be
declared in the PropertyCheck's options.

bug: 23939594
Change-Id: Ic22ad55ea5d965b88cfe6f4d86b9706023eb5d47