Speed up SELinux host tests.
ctsv2 now allows us to tag cts-artifacts to place resources in the cts
"home dir."  Place sepolicy-analyze in this directory to avoid the need
to extract it from the CtsSecurityHostTests jar on each test
invocation.  This reduces runtime by approximately 95%.  Also clean up
unused SELinuxNeverallowRulesTest frame code.

Adjust SELinuxHostTest to reflect the new location of sepolicy-analyze,
and move most of the file-extraction and pulling to individual tests
from the setUp() method to the methods in which they are used.  This
also reduces runtime by ober 90%, even with some files still extracted
from the jar in selected tests.  Leave some jar extraction for now so as
not to pollute the cts out dir for minimal gains.

Bug: 21266225
Change-Id: Iee4a7c34e48287ba8c613660b715ca641c227b8b